added user and manager role , dynamic api route
